big ansible update
I got rid of the `ansible.sh` wrapper as I don't need it anymore.

I shifted package related roles into a `pkg` directory. I also learned about how to use `when` conditionals and applied that to updates and upgrades. Basically there's a `do_upgrade` variable that has a default value of "false". When that is false the update checker runs. If it is true the upgrade process runs. Since the variable can be overridden on the command line with the `-e` option, which takes precedence over all other variable declarations, a standard run will show if there are updates and list them out. Then a run with `-e "do_upgrade=true"` will run the actual upgrade. I'm also using the "safe" upgrade from `aptitude` since I want to make extra sure that the `kubelet` package which is on hold remains held back when upgrades are happening.

Signed-off-by: Sean Johnson <sean@ttys0.net>